third_party/blink/web_tests/external/wpt/screen-capture/getdisplaymedia-framerate.https-expected.txt

This is a testharness.js-based test.
FAIL getDisplayMedia() must adhere to frameRate if set assert_greater_than_equal: expected a number greater than or equal to 2.5 but got NaN
Harness: the test ran to completion.
